Fixed map corruption due to BgEvent having u8 hiddenItemId. Sleep talk is now Normal instead of Sound.

+++ include/global.fieldmap.h
@@ -107,7 +107,7 @@ struct BgEvent
         u8 *script;
         struct {
             u16 item;
-            u8 hiddenItemId;
+            u16 hiddenItemId;
         } hiddenItem;
         u32 secretBaseId;
     } bgUnion;
